#8.4 Protecting Video Routes
- 영상을 Edit와 Delete할 때, 다른 사용자가 접근하는 것을 막아주는 코드 작성
- videoDetail 페이지에 나오는 creator 링크 색 변경
- creator의 링크를 누르면 해당 creator가 올린 영상을 확인할 수 있게 함

+++ controllers/videoController.js
@@ -16,6 +16,7 @@ export const home = async (req, res) => {
 export const search = async (req, res) => {
   const { query: { term: searchingBy } } = req;
   let videos = [];
+
   try {
     videos = await Video.find({ title: { $regex: searchingBy, $options: "i" } });
   } catch (error) {
@@ -60,14 +61,20 @@ export const videoDetail = async (req, res) => {
 // Edit Video
 export const getEditVideo = async (req, res) => {
   const {
-    params: { id }
+    params: { id },
+    user: { _id: userId }
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (JSON.stringify(video.creator) !== JSON.stringify(userId)) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
-    res.redirect(routes.home);
+    console.log(error);
   }
+  res.redirect(routes.home);
 };
 
 export const postEditVideo = async (req, res) => {
@@ -87,10 +94,16 @@ export const postEditVideo = async (req, res) => {
 // Delete Video
 export const deleteVideo = async (req, res) => {
   const {
-    params: { id }
+    params: { id },
+    user: { _id: userId }
   } = req;
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (JSON.stringify(video.creator) !== JSON.stringify(userId)) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
